Sptint 6 2020-09-02

+++ posts/views.py
@@ -1,14 +1,15 @@
+from django.contrib.auth.decorators import login_required
 from django.core.paginator import Paginator
 from django.shortcuts import get_object_or_404, redirect, render
-from django.contrib.auth.decorators import login_required
 from django.views.decorators.cache import cache_page
-from .forms import PostForm, CommentForm
-from .models import Post, Group, User, Follow
+
+from .forms import CommentForm, PostForm
+from .models import Follow, Group, Post, User
 
 
 @cache_page(20, key_prefix='index_page')
 def index(request):
-    post_list = Post.objects.all()
+    post_list = Post.objects.select_related('group').all()
     paginator = Paginator(post_list, 10)
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
@@ -50,16 +51,18 @@ def profile(request, username):
     user = request.user
     author = get_object_or_404(User, username=username)
     posts = author.posts.all()
-    post = author.posts.first()
-    posts_count = author.posts.count()
+    post = posts.first()
+    posts_count = posts.count()
     paginator = Paginator(posts, 10)
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
     following = Follow.objects.filter(
         user__username=user, author__username=author
         ).exists()
-    count_followers = Follow.objects.filter(author__username=author).count
-    count_following = Follow.objects.filter(user__username=author).count
+    # count_followers = Follow.objects.filter(author__useame=author).count
+    count_followers = author.following.count()
+    count_following = author.follower.count()
+
     context = {
         'author': author,
         'user': user,
@@ -134,25 +137,24 @@ def server_error(request):
 @login_required
 def add_comment(request, username, post_id):
     post = get_object_or_404(Post, pk=post_id, author__username=username)
-    if request.method == 'POST':
-        form = CommentForm(request.POST)
-        if form.is_valid():
-            comment = form.save(commit=False)
-            comment.author = request.user
-            comment.post = post
-            comment.save()
-            return redirect('post', username=username, post_id=post_id)
-    form = CommentForm()
-    return render(request, 'posts/comments.html', {'form': form, 'post': post})
+    form = CommentForm(request.POST or None)
+    if form.is_valid():
+        comment = form.save(commit=False)
+        comment.author = request.user
+        comment.post = post
+        comment.save()
+        return redirect('post', username=username, post_id=post_id)
+    return render(
+        request,
+        'includes/comments.html',
+        {'form': form, 'post': post}
+    )
 
 
 @login_required
 def follow_index(request):
     user = request.user
-    followed_users = User.objects.filter(following__user=request.user)
-    posts = []
-    for user in followed_users:
-        posts.extend(Post.objects.filter(author=user))
+    posts = Post.objects.filter(author__following__user=user)
     paginator = Paginator(posts, 10)
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
@@ -167,12 +169,10 @@ def follow_index(request):
 def profile_follow(request, username):
     user = request.user
     author = get_object_or_404(User, username=username)
-    follow_exist = Follow.objects.filter(
-        user__username=user, author__username=author
-        ).exists()
-    same_user = (user.pk == author.pk)
+    follow_exist = Follow.objects.filter(user=user, author=author).exists()
+    same_user = (user == author)
     if follow_exist or same_user:
-        return redirect('profile', username=author.username)
+        return redirect('profile', username=username)
     Follow.objects.create(user=user, author=author)
     return redirect('profile', username=author.username)
 
@@ -181,7 +181,5 @@ def profile_follow(request, username):
 def profile_unfollow(request, username):
     user = request.user
     author = get_object_or_404(User, username=username)
-    Follow.objects.filter(
-        user__username=user, author__username=author
-        ).delete()
+    Follow.objects.filter(user=user, author=author).delete()
     return redirect('profile', username=author.username)
